requirements.txt
watchdog
requests
yagmail
python-dotenv